backend/core/config.py
from dotenv import load_dotenv
load_dotenv()  # Load .env

import os
import yaml
import logging.config
from pydantic_settings import BaseSettings

class Settings(BaseSettings):
    # Project settings
    PROJECT_NAME: str = "Backend API Docs"
    PROJECT_VERSION: str = "1.0.0"
    PROJECT_DESCRIPTION: str = "Backend API Docs"    

    # Database settings
    DATABASE_URL: str
    DATABASE_URL_TEST: str
    DB_POOL_SIZE: int = 10
    DB_MAX_OVERFLOW: int = 20
    DB_POOL_TIMEOUT: int = 30
    DB_POOL_RECYCLE: int = 3600
    DB_CONNECT_TIMEOUT: int = 60
    DB_READ_TIMEOUT: int = 30
    DB_WRITE_TIMEOUT: int = 30

    # Redis settings
    REDIS_URL: str

    # CORS settings
    HOSTNAME: str
    BACKEND_PORT: str
    FRONTEND_PORT: str

    # JWT settings
    SECRET_KEY: str
    ALGORITHM: str = "RS256"
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 30

    # Logging settings
    LOG_LEVEL: str = "INFO"

    # Other settings
    DEBUG: bool = True
    SSL_ENABLE: bool = True
    RATE_LIMIT: int = 30
    RATE_LIMIT_WINDOW_SECONDS: int = 180  # 3 minutes
    BLOCK_TIME_SECONDS: int = 600  # 10 minutes
    
    # Manual settings
    UPLOAD_DIR: str = "uploads/manuals"
    MAX_FILE_SIZE: int = 1024 * 1024 * 1024  # 1GB
    
    # Keycloak settings
    KEYCLOAK_SERVER_URL: str
    KEYCLOAK_REALM: str
    KEYCLOAK_CLIENT: str
    KEYCLOAK_ADMIN_CLIENT: str
    KEYCLOAK_ADMIN_CLIENT_SECRET: str
    KEYCLOAK_VERIFY: bool = False
    KEYCLOAK_SUPER_ROLE: str = "tsadmin"

    # Web push settings
    VAPID_PUBLIC_KEY: str
    VAPID_PRIVATE_KEY: str
    VAPID_EMAIL: str

    # InfluxDB settings
    INFLUXDB_URL: str
    INFLUXDB_TOKEN: str
    INFLUXDB_ORG: str
    INFLUXDB_BUCKET: str
    INFLUXDB_TIMEOUT: int = 10

# Create a settings instance to be imported elsewhere
settings = Settings()

def setup_logging(yaml_path="logging_config.yaml"):
    os.makedirs("logs", exist_ok=True)
    with open(yaml_path, "r") as f:
        config = yaml.safe_load(f)
    # Override the root logger or specified logger's level with LOG_LEVEL from environment
    log_level = settings.LOG_LEVEL
    if "root" in config:
        config["root"]["level"] = log_level
    # If there are multiple loggers, override their levels as well
    if "loggers" in config:
        for logger in config["loggers"].values():
            logger["level"] = log_level
    logging.config.dictConfig(config)